GH-31 extract connector for sonar server
